CIV-640 move api path from civil citizen ui urls to its own constants file for civil service api calls
