+++ src/main/app/client/civilServiceUrls.ts
@@ -1,0 +1,1 @@
+export const CASES_URL = '/cases/';
