+++ src/main/routes/urls.ts
@@ -4,7 +4,6 @@ export const CALLBACK_URL: Urls= '/oauth2/callback';
 export const SIGN_IN_URL: Urls = '/login';
 export const SIGN_OUT_URL: Urls = '/logout';
 export const DASHBOARD_URL: Urls = '/dashboard';
-export const CASES_URL: Urls = '/cases/';
 export const CITIZEN_PHONE_NUMBER_URL: Urls = '/citizen-phone';
 export const ROOT_URL: Urls = '/home';
 export const DOB_URL: Urls = '/your-dob';
